+++ packages/connectivity/lib/connectivity.dart
@@ -14,10 +14,10 @@ import 'package:flutter/services.dart';
 enum ConnectivityResult { wifi, mobile, none }
 
 const MethodChannel _methodChannel =
-    const MethodChannel('plugins.flutter.io/connectivity');
+    MethodChannel('plugins.flutter.io/connectivity');
 
 const EventChannel _eventChannel =
-    const EventChannel('plugins.flutter.io/connectivity_status');
+    EventChannel('plugins.flutter.io/connectivity_status');
 
 class Connectivity {
   Stream<ConnectivityResult> _onConnectivityChanged;
@@ -41,6 +41,20 @@ class Connectivity {
   Future<ConnectivityResult> checkConnectivity() async {
     final String result = await _methodChannel.invokeMethod('check');
     return _parseConnectivityResult(result);
+  }
+
+  /// Obtains the wifi name (SSID) of the connected network
+  ///
+  /// Please note that it DOESN'T WORK on emulators (returns null).
+  ///
+  /// From android 8.0 onwards the GPS must be ON (high accuracy)
+  /// in order to be able to obtain the SSID.
+  Future<String> getWifiName() async {
+    String wifiName = await _methodChannel.invokeMethod('wifiName');
+    // as Android might return <unknown ssid>, uniforming result
+    // our iOS implementation will return null
+    if (wifiName == '<unknown ssid>') wifiName = null;
+    return wifiName;
   }
 }