+++ packages/firebase_messaging/test/firebase_messaging_test.dart
@@ -51,16 +51,14 @@ void main() {
 
   test('configure', () {
     firebaseMessaging.configure();
-    verify(mockChannel.setMethodCallHandler(typed(any)));
+    verify(mockChannel.setMethodCallHandler(any));
     verify(mockChannel.invokeMethod('configure'));
   });
 
   test('incoming token', () async {
     firebaseMessaging.configure();
     final dynamic handler =
-        verify(mockChannel.setMethodCallHandler(typed(captureAny)))
-            .captured
-            .single;
+        verify(mockChannel.setMethodCallHandler(captureAny)).captured.single;
     final String token1 = 'I am a super secret token';
     final String token2 = 'I am the new token in town';
     Future<String> tokenFromStream = firebaseMessaging.onTokenRefresh.first;
@@ -79,9 +77,7 @@ void main() {
   test('incoming iOS settings', () async {
     firebaseMessaging.configure();
     final dynamic handler =
-        verify(mockChannel.setMethodCallHandler(typed(captureAny)))
-            .captured
-            .single;
+        verify(mockChannel.setMethodCallHandler(captureAny)).captured.single;
     IosNotificationSettings iosSettings = const IosNotificationSettings();
 
     Future<IosNotificationSettings> iosSettingsFromStream =
@@ -102,17 +98,15 @@ void main() {
     final Completer<dynamic> onLaunch = new Completer<dynamic>();
     final Completer<dynamic> onResume = new Completer<dynamic>();
 
-    firebaseMessaging.configure(onMessage: (dynamic m) {
+    firebaseMessaging.configure(onMessage: (dynamic m) async {
       onMessage.complete(m);
-    }, onLaunch: (dynamic m) {
+    }, onLaunch: (dynamic m) async {
       onLaunch.complete(m);
-    }, onResume: (dynamic m) {
+    }, onResume: (dynamic m) async {
       onResume.complete(m);
     });
     final dynamic handler =
-        verify(mockChannel.setMethodCallHandler(typed(captureAny)))
-            .captured
-            .single;
+        verify(mockChannel.setMethodCallHandler(captureAny)).captured.single;
 
     final Map<String, dynamic> onMessageMessage = <String, dynamic>{};
     final Map<String, dynamic> onLaunchMessage = <String, dynamic>{};