+++ packages/firebase_ml_vision/lib/src/firebase_vision.dart
@@ -17,19 +17,39 @@ class FirebaseVision {
 
   @visibleForTesting
   static const MethodChannel channel =
-      const MethodChannel('plugins.flutter.io/firebase_ml_vision');
+      MethodChannel('plugins.flutter.io/firebase_ml_vision');
 
   /// Singleton of [FirebaseVision].
   ///
   /// Use this get an instance of a detector:
   ///
   /// ```dart
-  /// TextDetector textDetector = FirebaseVision.instance.getTextDetector();
+  /// TextDetector textDetector = FirebaseVision.instance.textDetector();
   /// ```
   static final FirebaseVision instance = new FirebaseVision._();
 
-  /// Creates an instance of [TextDetector];
-  TextDetector getTextDetector() => new TextDetector._();
+  /// Creates an instance of [BarcodeDetector].
+  BarcodeDetector barcodeDetector([BarcodeDetectorOptions options]) {
+    return BarcodeDetector._(options ?? const BarcodeDetectorOptions());
+  }
+
+  /// Creates an instance of [FaceDetector].
+  FaceDetector faceDetector([FaceDetectorOptions options]) {
+    return FaceDetector._(options ?? const FaceDetectorOptions());
+  }
+
+  /// Creates an instance of [LabelDetector].
+  LabelDetector labelDetector([LabelDetectorOptions options]) {
+    return LabelDetector._(options ?? const LabelDetectorOptions());
+  }
+
+  /// Creates an instance of [LabelDetector].
+  CloudLabelDetector cloudLabelDetector([VisionCloudDetectorOptions options]) {
+    return CloudLabelDetector._(options ?? const VisionCloudDetectorOptions());
+  }
+
+  /// Creates an instance of [TextDetector].
+  TextDetector textDetector() => new TextDetector._();
 }
 
 /// Represents an image object used for both on-device and cloud API detectors.
@@ -38,18 +58,29 @@ class FirebaseVision {
 class FirebaseVisionImage {
   FirebaseVisionImage._(this.imageFile);
 
+  /// Construct a [FirebaseVisionImage] from a file.
   factory FirebaseVisionImage.fromFile(File imageFile) {
+    assert(imageFile != null);
     return FirebaseVisionImage._(imageFile);
   }
 
+  /// Construct a [FirebaseVisionImage] from a file path.
   factory FirebaseVisionImage.fromFilePath(String imagePath) {
+    assert(imagePath != null);
     return FirebaseVisionImage._(new File(imagePath));
   }
 
+  /// The file location of the image.
   final File imageFile;
 }
 
 /// Abstract class for detectors in [FirebaseVision] API.
 abstract class FirebaseVisionDetector {
+  /// Uses machine learning model to detect objects of interest in an image.
   Future<dynamic> detectInImage(FirebaseVisionImage visionImage);
+}
+
+String _enumToString(dynamic enumValue) {
+  final String enumString = enumValue.toString();
+  return enumString.substring(enumString.indexOf('.') + 1);
 }